Fold FE builders into domain modules
