+++ tests/cli/test_GenerateFEM.py
@@ -193,7 +193,7 @@ def test_dry_run_prints_commands_without_running(monkeypatch, capsys):
     assert calls == []
     assert (
         capsys.readouterr().out.strip()
-        == "OgoSpineCompressionFe density.nii.gz spine_mask.nii.gz "
+        == "ogoFEA-spine-builder density.nii.gz spine_mask.nii.gz "
         "--mask_threshold 6 --process_mask_threshold 5 --appendix L2 "
         "--quality_control False"
     )
